@@ -8,10 +8,34 @@ class testRobot : public Robot
 {
 
   public:
-    testRobot(Trajectory_Controller &controller) : Robot(controller) {}
+    testRobot() {
+
+    Robot_Config config = {
+        .hz = 50,
+        .motion_constraints =
+            {
+                .max_velocity = 1.0,
+                .max_acceleration = 0.5,
+                .max_deceleration = 0.5,
+                .max_jerk = 0.0,
+            },
+
+    };
+
+    PIDGains linear_gains = {1.0, 0.0, 0.0};
+    PIDController linear_pid(linear_gains);
+    linear_pid.output_max = 10, 0;
+    linear_pid.output_min = 0, 0;
+    PIDGains angular_gains = {1.0, 0.0, 0.0};
+    PIDController angular_pid(angular_gains);
+    angular_pid.output_max = 10, 0;
+    angular_pid.output_min = 0, 0;
+	trajectory_controller = std::make_unique<Linear_Controller>(linear_pid, angular_pid, config.hz);
+
+
+
+	}
     void send_velocity_command(Velocity2d &cmd) override{};
-    void update_state() override{};
-    void read_sensors() override{};
     Pose_State read_state() override { return {}; };
 };
 
@@ -26,35 +50,17 @@ int main()
     std::vector<Ecef_Coord> waypoints = {
         {0.0, 0.0, 0.0}, {4.0, -2.0, 0.0}, {4.0, 2.0, 0.0}, {0.0, 0.0, 0.0}};
 
-    Robot_Config config = {
-        .hz = 50,
-        .motion_constraints =
-            {
-                .max_velocity = 1.0,
-                .max_acceleration = 0.5,
-                .max_deceleration = 0.5,
-                .max_jerk = 0.0,
-            },
-
-    };
-    PIDGains linear_gains = {1.0, 0.0, 0.0};
-    PIDController linear_pid(linear_gains);
-    linear_pid.output_max = 10, 0;
-    linear_pid.output_min = 0, 0;
-    PIDGains angular_gains = {1.0, 0.0, 0.0};
-    PIDController angular_pid(angular_gains);
-    angular_pid.output_max = 10, 0;
-    angular_pid.output_min = 0, 0;
-    Linear_Controller controller(config, linear_pid, angular_pid, config.hz);
 
     Ecef_Coord current = waypoints[0];
     Ecef_Coord next = waypoints[1];
 
-    testRobot robot(controller);
+    testRobot robot;
 
-    std::function<void()> bound_path_loop =
-        std::bind(&Linear_Controller::path_loop, &controller, std::ref(robot.path_queue),
-                  std::ref(waypoints));
+
+    std::function<void()> bound_path_loop = [&robot, &waypoints]() {
+		static_cast<Linear_Controller*>(robot.trajectory_controller.get())->path_loop(waypoints);
+	
+	};
     std::function<void()> bound_control_loop = std::bind(&testRobot::control_loop, &robot);
     std::thread path_loop = std::thread(worker_function, bound_path_loop, 30);
     std::thread control_loop_thread = std::thread(worker_function, bound_control_loop, 2);
