@@ -1,21 +1,20 @@
 #include "robot.hpp"
+#include "trajectory_controller.hpp"
 #include "types.hpp"
 
 void Robot::control_loop()
 {
-    update_state();
     pose_state = read_state();
     logger.savePosesToFile(frame_controller);
     logger.saveTimesToFile(motion_time);
     // Path_Movement path = readPath();
     // Sensors sensors = readSensors();
-    sensor_manager.readSensors();
+    // sensor_manager.readSensors();
     /*sensor_manager.sm_thread.join();*/
 
-    Velocity2d cmd = trajectory_controller.get_cmd(frame_controller, path_queue);
+    Velocity2d cmd = trajectory_controller->get_cmd(frame_controller, path_queue);
     send_velocity_command(cmd);
     /*motion_time += 1000 / hz;*/
-    motion_time += 1 / hz;
+    motion_time += 1.0 / hz;
 }
 
-void Robot::read_path() {}
